Improve server ping list response
Change the read timeout and connection timeout both to 5s. This
is to decrease the time to label servers as offline.

Additionally, increase the worker count for pinging servers to 128
from 5. This allows many more server pings to occur at the same time,
which allows users to scroll their serverlist.