Only tick entity tracker if players are tracking or if at entity ticking
Vanilla will only tick if the entity moves or is at entity ticking
level. For Paper, I am concerned that using exact Vanilla behavior
may cause problems with plugins - so we will settle on ticking the
tracker as well if players are tracking the entity.

This should reduce ticking for larger VD settings.