Don't call recalcBlockCounts on the client
This logic is expensive, and we only need to know if there may be any special colliding blocks in the section.

Before this change, ClientboundLevelChunkWithLightPacket was the top contributor to frame time spikes, after this change it's nowhere near the top of the profile. Specifically, the recalc logic was taking upwards of 80% of the processing time for this packet, the simplified check only takes <5%.