Update to 1.21.2-pre1
For collisions, need to check out the new applyEffectsFromBlocks
function and see if there are any improvements we can make.

For the chunk system, we need to implement async chunk saving.
The current async unload structure will not work, as it is designed
to only handle 1 pending save at any given time.