Improve server tick loop
The Vanilla tick loop works as follows:

1. Update the next tick start time
   Note: The next tick start time may be in the past, in which case
         the server is sprinting to perform catchup
         which seems to be ~20-100 ticks on first glance
2. Execute the tick
3. Process all player packets
4. Process all tick tasks if the following conditions are met:
   1. There is still time until the next tick
      The next tick may be delayed (up to end tick time + tick interval)
      if there are still tasks in queue that are processable (2)
   2. The next tick task was queued at least 3 ticks ago

The new tick loop works as follows:

1. Update the next tick start time
   Note: Depending on the new tick catchup config value, the
         next tick start time may be in the past by up to the
         configured value. The catchup may be disabled.
2. Execute all tasks
   Note: "before the start of the tick" refers to any time before (1)
   1. All tick tasks queued before the start of the tick are executed
   2. All pending player packet tasks are executed
   3. All chunk system tasks queued before the start of the tick are executed
3. Execute the tick
4. Process tick tasks and player packet tasks until the start of the next tick

There are three key differences in the tick loops described above:
1. Player packet tasks are processed while waiting for the next tick
2. The new tick loop ensures that all tasks are processed each tick
3. Tick catchup is configurable and may be disabled

Player packet tasks are important to process inbetween ticks so that
average packet processing latency is lower. If a server is completely
idle, it should be able to process packets with a latency close to
the wakeup time - rather than an average latency of the tick time / 2
(25ms for 20tps). This should in general reduce perceived latency
for connected players. This change did not need to be done until
1.21.9 where Mojang moved packet processing off of the tick tasks.

Processing all other tasks should additionally help the chunk
system keep up with chunk loading requests, regardless of
the server load.

By default, tick catchup is lowered to a max of 5 ticks. This allows
some catchup to occur, but should greatly limit the existing speedup
which disrupts players (i.e mobs attacking and moving faster) after
lagspikes occur.

Note that since all task processing is moved into the server tick,
this will cause a higher reported MSPT value. However, this is
a result of more properly timing - so I have no plans to adjust it.