Fix Entity#updateFluidHeightAndDoFluidPushing inconsistency with Vanilla
The first issue was that the maxHeightDiff was not used to scale
the fluid push vector. Fixing the first issue revealed that the
second issue was that the result of the push vector is
order-dependent on the blocks iterated. The second issue was resolved
by instead caching the chunks in a local array (so that the Vanilla
loop order could be done) rather than iterating over each chunk
individually.